Extract `CollectionUpdateOperations` into separate `shard` crate (#7054)
