+++ lib/shard/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name = "shard"
+version = "0.1.0"
+authors = [
+    "Andrey Vasnetsov <vasnetsov93@gmail.com>",
+    "Qdrant Team <info@qdrant.tech>",
+]
+license = "Apache-2.0"
+edition = "2024"
+
+[dependencies]
+api = { path = "../api" }
+common = { path = "../common/common" }
+segment = { path = "../segment" }
+sparse = { path = "../sparse" }
+
+ahash = { workspace = true }
+itertools = { workspace = true }
+rand = { workspace = true }
+schemars = { workspace = true }
+serde = { workspace = true }
+strum = { workspace = true }
+tonic = { workspace = true }
+validator = { workspace = true }
+
+[dev-dependencies]
+proptest = { workspace = true }
+serde_cbor = { workspace = true }
+serde_json = { workspace = true }
