+++ lib/shard/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod operations;
+
+pub type PeerId = u64;
