+++ src/common/update.rs
@@ -4,19 +4,19 @@ use api::rest::models::InferenceUsage;
 use api::rest::*;
 use collection::collection::Collection;
 use collection::operations::conversions::write_ordering_from_proto;
-use collection::operations::payload_ops::*;
 use collection::operations::point_ops::*;
 use collection::operations::shard_selector_internal::ShardSelectorInternal;
 use collection::operations::types::{CollectionResult, UpdateResult};
 use collection::operations::vector_ops::*;
 use collection::operations::verification::*;
-use collection::operations::*;
 use collection::shards::shard::ShardId;
 use common::counter::hardware_accumulator::HwMeasurementAcc;
 use schemars::JsonSchema;
 use segment::json_path::JsonPath;
 use segment::types::{PayloadFieldSchema, PayloadKeyType, StrictModeConfig};
 use serde::{Deserialize, Serialize};
+use shard::operations::payload_ops::*;
+use shard::operations::*;
 use storage::content_manager::collection_meta_ops::*;
 use storage::content_manager::errors::StorageError;
 use storage::content_manager::toc::TableOfContent;
